Pipeline and services refactor 6: health and utils
